requirements.txt
django==3.0.3
redis==3.5.3
python-telegram-bot==12.8
gunicorn==20.0.4
